Profile Page - Successful data view

+++ app/src/main/java/com/louie/warzonestats/models/player/PropertiesXXXXXXXXXXXX.kt
@@ -2,11 +2,11 @@ package com.louie.warzonestats.models.player
 
 data class PropertiesXXXXXXXXXXXX(
     val deaths: Int,
-    val kdRatio: Int,
+    val kdRatio: Double,
     val kills: Int,
     val plants: Int,
     val revives: Int,
     val score: Int,
-    val scorePerMinute: Int,
+    val scorePerMinute: Double,
     val timePlayed: Int
 )
