+++ app/src/main/java/com/louie/warzonestats/models/player/PropertiesXXXXXXXXXXXXXXX.kt
@@ -3,10 +3,10 @@ package com.louie.warzonestats.models.player
 data class PropertiesXXXXXXXXXXXXXXX(
     val deaths: Int,
     val defuses: Int,
-    val kdRatio: Int,
+    val kdRatio: Double,
     val kills: Int,
     val plants: Int,
     val score: Int,
-    val scorePerMinute: Int,
+    val scorePerMinute: Double,
     val timePlayed: Int
 )
