+++ app/src/main/java/com/louie/warzonestats/models/player/PropertiesXXXXXXXXXXXXXXXXXX.kt
@@ -3,10 +3,10 @@ package com.louie.warzonestats.models.player
 data class PropertiesXXXXXXXXXXXXXXXXXX(
     val deaths: Int,
     val infected: Int,
-    val kdRatio: Int,
+    val kdRatio: Double,
     val kills: Int,
     val score: Int,
-    val scorePerMinute: Int,
+    val scorePerMinute: Double,
     val time: Int,
     val timePlayed: Int
 )
